+++ src/Solution10/Collection/Collection.php
@@ -78,9 +78,9 @@ class Collection implements \Countable, \ArrayAccess, \Iterator
      */
     protected function addBaseSelectors()
     {
-        $this->addSelector('(?P<start>-?[0-9]+):(?P<end>[0-9]+|END)', array($this, 'splice'));
+        $this->addSelector('(?P<start>-?[0-9]+):(?P<end>[0-9]+|END)', array($this, 'spliceCallback'));
+        $this->addSelector('[0-9a-zA-Z\-_]+,', array($this, 'pluckCallback'));
     }
-
 
     /**
      * Calls a selector or throws an exception if it doesn't know what it is.
@@ -94,11 +94,94 @@ class Collection implements \Countable, \ArrayAccess, \Iterator
         foreach ($this->selectors as $selector => $callback) {
             $regex = '/' . $selector . '/i';
             if (preg_match($regex, $key, $matches)) {
-                return call_user_func_array($callback, array($this, $matches));
+                return call_user_func_array($callback, array($this, $key, $matches));
             }
         }
 
         throw new Exception\Index('Unknown index: ' . $key);
+    }
+
+    /*
+     * ------------------------ Base Selectors --------------------
+     */
+
+    /**
+     * Splices a collection given by a range. You can also use START and END as
+     * keywords.
+     *
+     * Splicing is 1 indexed!!!
+     *
+     *  $result = $collection['1:10']; // Get first 10 items.
+     *
+     * @param   Collection  $collection     Collection we're operating on.
+     * @param   string      $key            The unadultarated key the user passed.
+     * @param   array       $matches        regex matches array
+     * @return  array
+     * @throws  Exception\Bounds
+     */
+    public function spliceCallback(Collection $collection, $key, array $matches)
+    {
+        // $collection isn't used as this is an internal function.
+        $start = $matches['start'];
+        $end = $matches['end'];
+
+        // You can use the END keyword to select everything up until an end point.
+        if ($end == 'END') {
+            $end = count($this->contents) - 1;
+        }
+
+        // If the start is negative, we count backwards from the end of the CSV. array_slice can handle negative
+        // offsets, but bounds checking gets a bit gnarly.
+        if ($start < 0) {
+            $start = (count($this->contents)) - abs($start);
+        }
+
+        // Check the bounds:
+        if ($start >= count($this->contents)) {
+            throw new Exception\Bounds(
+                'Start index (' . $start . ') is beyond the end of the file',
+                Exception\Bounds::ERROR_START_OUT_OF_RANGE
+            );
+        } elseif ($start > $end) {
+            throw new Exception\Bounds(
+                'Start index is greater than end index: ' . $start . ' > ' . $end,
+                Exception\Bounds::ERROR_START_GT_END
+            );
+        }
+
+        // If the end index is > the total, we set to the total:
+        if ($end >= count($this->contents)) {
+            $end = count($this->contents) - 1;
+        }
+
+        // Now work out the params for array_slice
+        $offset = $start;
+        $length = ($end - $start) + 1;
+
+        return array_slice($this->contents, $offset, $length);
+    }
+
+    /**
+     * Implements the plucking behaviour.
+     *
+     * @param   Collection  $collection     Collection we're operating on.
+     * @param   string      $key            The unadultarated key the user passed.
+     * @param   array       $matches        regex matches array
+     * @return  array
+     */
+    public function pluckCallback(Collection $collection, $key, array $matches)
+    {
+        $indexes = explode(',', $key);
+        $result = array();
+        foreach ($indexes as $idx) {
+            $idx = trim($idx);
+            if ($idx != '') {
+                if (array_key_exists($idx, $this->contents)) {
+                    $result[$idx] = $this->contents[$idx];
+                }
+            }
+        }
+        return $result;
     }
 
     /*
@@ -128,7 +211,7 @@ class Collection implements \Countable, \ArrayAccess, \Iterator
 
 
     /**
-     * This function contains a lot of the magic with splicing and such.
+     * This function is the gateway to the short syntax magic.
      *
      * @param   mixed   $offset     INT for numeric index. String splice selector otherwise.
      * @throws  Exception\Index
@@ -140,46 +223,8 @@ class Collection implements \Countable, \ArrayAccess, \Iterator
         if (is_int($offset)) {
             return $this->contents[$offset];
         } elseif ($offset == ':last') {
-            // Shortcut for fetching the last item in the CSV:
+            // Shortcut for fetching the last item in the Collection:
             return $this->contents[count($this->contents) - 1];
-        } elseif (preg_match('/(?P<start>-?[0-9]+):(?P<end>[0-9]+|END)/', $offset, $matches)) {
-            $start = $matches['start'];
-            $end = $matches['end'];
-
-            // You can use the END keyword to select everything up until an end point.
-            if ($end == 'END') {
-                $end = count($this->contents) - 1;
-            }
-
-            // If the start is negative, we count backwards from the end of the CSV. array_slice can handle negative
-            // offsets, but bounds checking gets a bit gnarly.
-            if ($start < 0) {
-                $start = (count($this->contents)) - abs($start);
-            }
-
-            // Check the bounds:
-            if ($start >= count($this->contents)) {
-                throw new Exception\Bounds(
-                    'Start index (' . $start . ') is beyond the end of the file',
-                    Exception\Bounds::ERROR_START_OUT_OF_RANGE
-                );
-            } elseif ($start > $end) {
-                throw new Exception\Bounds(
-                    'Start index is greater than end index: ' . $start . ' > ' . $end,
-                    Exception\Bounds::ERROR_START_GT_END
-                );
-            }
-
-            // If the end index is > the total, we set to the total:
-            if ($end >= count($this->contents)) {
-                $end = count($this->contents) - 1;
-            }
-
-            // Now work out the params for array_slice
-            $offset = $start;
-            $length = ($end - $start) + 1;
-
-            return array_slice($this->contents, $offset, $length);
         } else {
             return $this->callSelector($offset);
         }
@@ -473,5 +518,56 @@ class Collection implements \Countable, \ArrayAccess, \Iterator
     public function cycleValue()
     {
         return $this[$this->cyclePosition];
+    }
+
+    /*
+     * --------------- Searching Functionality ----------------
+     */
+
+    /**
+     * Searching is dead simple. If you don't provide a callback, then the search
+     * will go through and check $term === $item. If you need something more powerful,
+     * you can provide a function to filter by:
+     *
+     *  function($term, $item) {
+     *      // Returning TRUE means that yes this matches the search. FALSE means no.
+     *  }
+     *
+     * The original keys from the collection are preserved! So if you match items in
+     * index 0 and 3, your array will be array(0 => value, 3 => value)
+     *
+     * @param   mixed           $term       Search term
+     * @param   null|callable   $callback   Search function to use as a filter.
+     * @return  array
+     * @throws  Exception\Search    If the callback is invalid.
+     */
+    public function search($term, $callback = null)
+    {
+        if ($callback !== null && !is_callable($callback)) {
+            throw new Exception\Search('Bad callback passed to search', Exception\Search::BAD_CALLBACK);
+        }
+
+        $result = array();
+        foreach ($this->contents as $key => $value) {
+            if ($callback === null) {
+                $include = ($value === $term);
+            } else {
+                $include = call_user_func_array($callback, array($term, $value));
+
+                // Check that the callback came back with a boolean:
+                if (!is_bool($include)) {
+                    throw new Exception\Search(
+                        'Bad return value from callback: '.$include,
+                        Exception\Search::BAD_CALLBACK_RETURN
+                    );
+                }
+            }
+
+            if ($include) {
+                $result[$key] = $value;
+            }
+        }
+
+        return $result;
     }
 }